remove byte cast

+++ src/main/java/org/codelibs/elasticsearch/vector/index/query/ScoreScriptUtils.java
@@ -238,7 +238,7 @@ public class ScoreScriptUtils {
                 if (value != 0) {
                     value = 1;
                 }
-                buf[offset] |= (byte) (value << pos);
+                buf[offset] |= value << pos;
                 pos--;
                 if (pos < 0) {
                     offset++;
